remove credential validation from repositories, and add tags to article view
