+++ build.rs
@@ -1,1 +1,3 @@
-fn main() {}
+fn main() {
+    println!("cargo:rerun-if-changed=templates/**/*.mk");
+}
