+++ web/static/css/layout.css
@@ -366,3 +366,17 @@ body {
 .masthead {
   font-size: 2.5em;
 }
+
+.tags {
+  display: flex;
+  gap: 5px;
+}
+
+.tag-link {
+  background: var(--color-bg-secondary);
+  color: var(--color-text-muted);
+  border-radius: 3px;
+  border: 1px solid var(--color-border-secondary);
+  padding: 0px 10px;
+  font-size: 0.7em;
+}
